requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==4.9.3